// src/enums/TechnologyTypes.ts
enum TechnologyTypes {
  html = "HTML",
  css = "CSS",
  jquery = "jQuery",
  javascript = "JavaScript",
  react = "React",
  reduxtoolkit = "Redux Toolkit",
  nextjs = "Next.js",
  rails = "Ruby on Rails",
  firebase = "Firebase",
  rspec = "RSpec",
  heroku = "Heroku",
  bootstrap = "Bootstrap",
  tailwindcss = "Tailwind CSS",
  vercel = "Vercel",
}

export default TechnologyTypes;